requirements/local.txt
-r base.txt

# testing, linting, and formatting tools
pytest==7.4.2
pytest-django==4.5.2

factory-boy==3.3.0
Faker==19.6.1

ipdb==0.13.13
ipython==8.15.0

mypy==1.5.1

django-stubs==4.2.4
djangorestframework-stubs==3.14.2

flake8==6.1.0
isort==5.12.0
black==23.9.1
pre-commit==3.4.0